huggingface/requirements.txt
torch>=2.0.0
transformers>=4.35.0
pillow>=9.0.0
huggingface_hub>=0.19.0
gradio>=4.0.0
timm>=0.9.0
